Break classes into their own file.
